Small additions

+++ models/writer.h
@@ -51,7 +51,7 @@ namespace ug{
 					
 					for (int i=0;i<points_y;i++){
 						for (int j=0;j<points_x;j++){
-							output<<pos[i*points_x+j]<<"\t";
+							output<<(double)pos[i*points_x+j]<<"\t";
 						}
 						output<<"\n";
 					}	
@@ -71,7 +71,7 @@ namespace ug{
 					for (int j=0;j<points_x;j++){
 						F wx=(j/(points_x-1.0))*dimX;
 						F wy=(F(1.0)-(i/(points_y-1.0)))*dimY;
-						output<<"("<<wy<<","<<wx<<")\t";
+						output<<"("<<(double)wy<<","<<(double)wx<<")\t";
 					}
 					output<<"\n";
 				}
